feat(ui): unify dual notification centers; feat(demo-seed): nordwind IR seeding (GAP-16); fix(api): markAsRead integer id
- GlobalNotificationCenter becomes the single notification surface: DB-backed notifications via tRPC merged with system notifications, deep-link resolution extracted to pure helper notificationDestinations.ts (:clientId/:id substitution, external http(s) new-tab with noopener,noreferrer), type-to-dismiss + outside-click close, aria/keyboard a11y, UI-STANDARD token-only surfaces

+++ packages/core/src/server/routers/notifications.ts
@@ -131,7 +131,7 @@ export const createNotificationsRouter = (t: any, clientProcedure: any, adminPro
             }),
 
         markAsRead: clientProcedure
-            .input(z.object({ id: z.number() }))
+            .input(z.object({ id: z.number().int() }))
             .mutation(async ({ input, ctx }: any) => {
                 const dbConn = await db.getDb();
                 await dbConn.update(schema.notificationLog)
